+++ collectionexercisesvc/src/main/java/uk/gov/ons/ctp/response/collection/exercise/repository/CollectionExerciseRepository.java
@@ -1,7 +1,5 @@
 package uk.gov.ons.ctp.response.collection.exercise.repository;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import uk.gov.ons.ctp.response.collection.exercise.domain.CollectionExercise;
@@ -10,6 +8,6 @@ import uk.gov.ons.ctp.response.collection.exercise.domain.CollectionExercise;
  * Spring JPA Repository for Collection Exercise
  *
  */
-public interface CollectionExerciseRepository extends JpaRepository<CollectionExercise, BigInteger> {
+public interface CollectionExerciseRepository extends JpaRepository<CollectionExercise, Integer> {
 
 }
