+++ collectionexercisesvc/src/main/java/uk/gov/ons/ctp/response/collection/exercise/repository/SampleUnitGroupRepository.java
@@ -1,10 +1,6 @@
 package uk.gov.ons.ctp.response.collection.exercise.repository;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import uk.gov.ons.ctp.response.collection.exercise.domain.ExerciseSampleUnitGroup;
@@ -14,7 +10,6 @@ import uk.gov.ons.ctp.response.collection.exercise.domain.ExerciseSampleUnitGrou
  *
  */
 @Repository
-public interface SampleUnitGroupRepository extends JpaRepository<ExerciseSampleUnitGroup, BigInteger> {
-
+public interface SampleUnitGroupRepository extends JpaRepository<ExerciseSampleUnitGroup, Integer> {
 
 }
